requirements.txt
# document loaders
pymupdf
python-docx
beautifulsoup4
requests

# langchain core
langchain
langchain-community
langchain-google-genai
google-generativeai

# vector stores
faiss-cpu
chromadb

# embeddings
sentence-transformers

# chunking / NLP
nltk

# config & utils
pyyaml
pydantic
python-dotenv

# benchmarking
pandas
numpy

# testing
pytest
pytest-cov